Moved PaimaPrimitive & Registry to paima/sm package
