+++ e2e/shared/my-primitives/src/mod.ts
@@ -1,10 +1,4 @@
-
-
-export * from "./PaimaPrimitive.ts";
-export * from "./PrimitiveRegistry.ts";
-
-
 export * from "./evm-erc721/mod.ts";
 export * from "./evm-erc20/mod.ts";
-export * from "./midnight-generic/mod.ts";
 export * from "./evm-paimal2/mod.ts";
+export * from "./midnight-generic/mod.ts";
