+++ packages/node-sdk/sm/mod.ts
@@ -1,5 +1,5 @@
 export * from "./PaimaSTM.ts";
-export * from "./primitives/processing.ts";
+export * from "./primitives/mod.ts";
 export * from "./types.ts";
 export * from "./delegate-wallet.ts";
 export * from "./verify-signature.ts";
