+++ packages/node-sdk/sm/primitives/mod.ts
@@ -1,0 +1,4 @@
+export * from "./processing.ts";
+export * from "./PaimaPrimitive.ts";
+export * from "./PrimitiveRegistry.ts";
+export * from "./types.ts";
