+++ packages/node-sdk/sm/primitives/types.ts
@@ -1,0 +1,14 @@
+/**
+ * Valid JSON Object, required for the db accounting.
+ */
+export type JsonValue =
+  | string
+  | number
+  | boolean
+  | null
+  | JsonValue[]
+  | { [key: string]: JsonValue };
+
+export type JsonObject = {
+  [key: string]: JsonValue;
+};
